wip commit so i dont lose dis hard work
