--- metrics-exporter-log/CHANGELOG.md
@@ -1,23 +1,0 @@
-# Changelog
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [Unreleased]
-
-## [0.3.0] - 2019-07-29
-### Changed
-- Switch to metrics-core 0.5.0.
-
-## [0.2.1] - 2019-06-11
-### Changed
-- Switch to metrics-core 0.4.0.
-
-## [0.2.0] - 2019-05-01
-### Changed
-- Switch to metrics-core 0.3.0.
-
-## [0.1.0] - 2019-04-23
-### Added
-- Effective birth of the crate.
