--- metrics-exporter-log/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "metrics-exporter-log"
-version = "0.4.0"
-authors = ["Toby Lawrence <toby@nuclearfurnace.com>"]
-edition = "2018"
-
-license = "MIT"
-
-description = "A metrics-core compatible exporter for forwarding metrics to logs."
-homepage = "https://github.com/metrics-rs/metrics"
-repository = "https://github.com/metrics-rs/metrics"
-documentation = "https://docs.rs/metrics-exporter-log"
-readme = "README.md"
-
-categories = ["development-tools::debugging"]
-keywords = ["metrics", "metrics-core", "exporter", "log"]
-
-[dependencies]
-metrics-core = { path = "../metrics-core", version = "^0.5" }
-log = "^0.4"
-tokio = { version = "0.2", features = ["time"] }
