--- metrics-exporter-log/README.md
@@ -1,18 +1,0 @@
-# metrics-exporter-log
-
-[![conduct-badge][]][conduct] [![downloads-badge][] ![release-badge][]][crate] [![docs-badge][]][docs] [![license-badge][]](#license)
-
-[conduct-badge]: https://img.shields.io/badge/%E2%9D%A4-code%20of%20conduct-blue.svg
-[downloads-badge]: https://img.shields.io/crates/d/metrics-exporter-log.svg
-[release-badge]: https://img.shields.io/crates/v/metrics-exporter-log.svg
-[license-badge]: https://img.shields.io/crates/l/metrics-exporter-log.svg
-[docs-badge]: https://docs.rs/metrics-exporter-log/badge.svg
-[conduct]: https://github.com/metrics-rs/metrics/blob/master/CODE_OF_CONDUCT.md
-[crate]: https://crates.io/crates/metrics-exporter-log
-[docs]: https://docs.rs/metrics-exporter-log
-
-__metrics-exporter-log__ is a metrics-core compatible exporter for forwarding metrics to logs.
-
-## code of conduct
-
-**NOTE**: All conversations and contributions to this project shall adhere to the [Code of Conduct][conduct].
