+++ metrics-macros/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "metrics-macros"
+version = "0.1.0"
+authors = ["Toby Lawrence <toby@nuclearfurnace.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+proc-macro = true
+
+[dependencies]
+metrics-core = { path = "../metrics-core", version = "^0.5" }
+syn = { version = "^1.0", features = ["extra-traits"] }
+quote = "^1.0"
+proc-macro2 = "^1.0"
+proc-macro-hack = "^0.5"
