--- metrics-observer-json/README.md
@@ -1,18 +1,0 @@
-# metrics-observer-json
-
-[![conduct-badge][]][conduct] [![downloads-badge][] ![release-badge][]][crate] [![docs-badge][]][docs] [![license-badge][]](#license)
-
-[conduct-badge]: https://img.shields.io/badge/%E2%9D%A4-code%20of%20conduct-blue.svg
-[downloads-badge]: https://img.shields.io/crates/d/metrics-observer-json.svg
-[release-badge]: https://img.shields.io/crates/v/metrics-observer-json.svg
-[license-badge]: https://img.shields.io/crates/l/metrics-observer-json.svg
-[docs-badge]: https://docs.rs/metrics-observer-json/badge.svg
-[conduct]: https://github.com/metrics-rs/metrics/blob/master/CODE_OF_CONDUCT.md
-[crate]: https://crates.io/crates/metrics-observer-json
-[docs]: https://docs.rs/metrics-observer-json
-
-__metrics-observer-json__ is a metrics-core compatible observer that outputs JSON.
-
-## code of conduct
-
-**NOTE**: All conversations and contributions to this project shall adhere to the [Code of Conduct][conduct].
