--- metrics-observer-yaml/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "metrics-observer-yaml"
-version = "0.1.1"
-authors = ["Toby Lawrence <toby@nuclearfurnace.com>"]
-edition = "2018"
-
-license = "MIT"
-
-description = "A metrics-core compatible observer that outputs YAML."
-homepage = "https://github.com/metrics-rs/metrics"
-repository = "https://github.com/metrics-rs/metrics"
-documentation = "https://docs.rs/metrics-observer-yaml"
-readme = "README.md"
-
-categories = ["development-tools::debugging"]
-keywords = ["metrics", "telemetry", "yaml"]
-
-[dependencies]
-metrics-core = { path = "../metrics-core", version = "^0.5" }
-metrics-util = { path = "../metrics-util", version = "^0.3" }
-hdrhistogram = { version = "^6.3", default-features = false }
-serde_yaml = "^0.8"
