--- metrics-observer-yaml/README.md
@@ -1,18 +1,0 @@
-# metrics-observer-yaml
-
-[![conduct-badge][]][conduct] [![downloads-badge][] ![release-badge][]][crate] [![docs-badge][]][docs] [![license-badge][]](#license)
-
-[conduct-badge]: https://img.shields.io/badge/%E2%9D%A4-code%20of%20conduct-blue.svg
-[downloads-badge]: https://img.shields.io/crates/d/metrics-observer-yaml.svg
-[release-badge]: https://img.shields.io/crates/v/metrics-observer-yaml.svg
-[license-badge]: https://img.shields.io/crates/l/metrics-observer-yaml.svg
-[docs-badge]: https://docs.rs/metrics-observer-yaml/badge.svg
-[conduct]: https://github.com/metrics-rs/metrics/blob/master/CODE_OF_CONDUCT.md
-[crate]: https://crates.io/crates/metrics-observer-yaml
-[docs]: https://docs.rs/metrics-observer-yaml
-
-__metrics-observer-yaml__ is a metrics-core compatible observer that outputs YAML.
-
-## code of conduct
-
-**NOTE**: All conversations and contributions to this project shall adhere to the [Code of Conduct][conduct].
