+++ metrics-runtime/Cargo.toml
@@ -15,11 +15,6 @@ readme = "README.md"
 categories = ["development-tools::debugging"]
 keywords = ["metrics", "telemetry", "histogram", "counter", "gauge"]
 
-[features]
-default = ["exporters", "observers"]
-exporters = ["metrics-exporter-log", "metrics-exporter-http"]
-observers = ["metrics-observer-yaml", "metrics-observer-json", "metrics-observer-prometheus"]
-
 [[bench]]
 name = "histogram"
 harness = false
@@ -33,11 +28,6 @@ arc-swap = "^0.3"
 parking_lot = "^0.9"
 quanta = "^0.3"
 crossbeam-utils = "^0.6"
-metrics-exporter-log = { path = "../metrics-exporter-log", version = "^0.4", optional = true }
-metrics-exporter-http = { path = "../metrics-exporter-http", version = "^0.3", optional = true }
-metrics-observer-yaml = { path = "../metrics-observer-yaml", version = "^0.1", optional = true }
-metrics-observer-prometheus = { path = "../metrics-observer-prometheus", version = "^0.1", optional = true }
-metrics-observer-json = { path = "../metrics-observer-json", version = "^0.1", optional = true }
 
 [dev-dependencies]
 log = "^0.4"
